+++ .github/workflows/check_new_mdfs.yml
@@ -16,6 +16,8 @@ jobs:
     steps:
       - name: Checkout repository
         uses: actions/checkout@v3
+        with:
+          ref: dev
 
       - name: Set up Python
         uses: actions/setup-python@v4
