+++ pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
     "click>=8.1.8",
     "liquichange>=0.2.1",
     "minicypher",
+    "packaging>=24.2",
     "prefect>=3.1.15",
     "pytest>=8.3.4",
 ]
